Directly invoke `make` during the container build
I don't know how the upstream build ever worked.

This change is intended to get GitHub Actions building, as the first
attempt failed because it couldn't find .build/linux-amd64. That
directory doesn't get created during a local build and doesn't appear to
be mentioned in any capacity whatsoever in this repository.