Add --mysqld.socket option
When DATA_SOURCE_NAME was removed it became impossible to configure
mysqld_exporter to connect to MySQL via a socket without resorting to a
my.cnf file.

This patch adds `--mysqld.socket` as an alternative to
`--mysqld.address` and accompaniment to `--mysqld.username`. We expect
it to be used in lieu of `--config.my-cnf`.